Fix multiple opens of same dataset not being recognized
Datasets don't have a consistent universal name,
so the dset table being indexed by dset name led to several
problems. One of these was bad object cleanup,
which resulted in file close not working properly. This
is the cause of the 'cork' test failing with the Bypass VOL.

We could theoretically try to use a different index into the dataset table,
but I think it makes the most sense to remove it outright and acquire
the necessary information locally during reads, which is the approach
this PR takes.